# coletor_go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Motor de Ingestão e Gestão Patrimonial.",
        "title": "QuantAdvisor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "100.95.28.45:8080",
    "basePath": "/api",
    "paths": {
...
        "/institucional/curva-capital": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna a evolução patrimonial e a volatilidade do mercado para o Tearsheet.",
                "tags": [
                    "6. Institucional (B2B)"
                ],
                "summary": "Curva de Capital",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Usuário",
                        "name": "usuario_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/institucional/replay": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os logs de auditoria imutáveis (Z-Score, Kelly, Fricção) do robô.",
                "tags": [
                    "6. Institucional (B2B)"
                ],
                "summary": "Replay de Decisão da IA",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Usuário",
                        "name": "usuario_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/institucional/resumo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna métricas B2B de performance (AUM, Sharpe, Drawdown).",
                "tags": [
                    "6. Institucional (B2B)"
                ],
                "summary": "Resumo Institucional (Tearsheet)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Usuário",
                        "name": "usuario_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/parametros": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna as configurações de Z-Score, Kelly e Travas de Risco do robô.",
                "tags": [
                    "8. Parâmetros e Configurações"
                ],
                "summary": "Obter Parâmetros de Risco",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Usuário",
                        "name": "usuario_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ParametrosOperacionais"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza a sensibilidade e os limites de segurança da Inteligência Artificial.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "8. Parâmetros e Configurações"
                ],
                "summary": "Salvar Parâmetros de Risco",
                "parameters": [
                    {
                        "description": "Parâmetros",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ParametrosOperacionais"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ParametrosOperacionais": {
            "type": "object",
            "properties": {
                "bloqueio_sentimento_negativo": {
                    "type": "boolean"
                },
                "custo_friccao_padrao": {
                    "type": "number"
                },
                "gatilho_rebalanceamento": {
                    "type": "number"
                },
                "limite_concentracao_ativo": {
                    "type": "number"
                },
                "modo_isencao_fiscal_estrita": {
                    "type": "boolean"
                },
                "multiplicador_kelly": {
                    "type": "number"
                },
                "multiplicador_stop_var": {
                    "type": "number"
                },
                "piso_max_drawdown": {
                    "type": "number"
                },
                "usuario_id": {
                    "type": "integer"
                },
                "z_score_compra_forte": {
                    "type": "number"
                },
                "z_score_stop_loss": {
                    "description": "👈 NOVO",
                    "type": "number"
                },
                "z_score_venda_lucro": {
                    "description": "👈 NOVO",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}